requirements.txt
snowflake-connector-python
pandas
openai
numpy
pyarrow
requests
dbt-snowflake
astronomer-cosmos==1.0.4
spotipy